Auction Payment

+++ Infrastructure/Ef/DbEntities/DbAuctionPayment.cs
@@ -4,9 +4,11 @@ namespace Infrastructure.Ef.DbEntities;
 
 public class DbAuctionPayment
 {
+    
     public int Id { get; set; }
+    
     [ForeignKey("IdUser")]
     public int IdUser { get; set; }
     public string Title { get; set; }
-    public float Price { get; set; }
+    public decimal Price { get; set; }
 }
